pyproject.toml
[tool.poetry]
# name = "poetry"
name = "lab_competitive_analysis"
version = "0.1.0"
description = ""
authors = ["Jaeyeop Chung <ge23gut@mytum.de>"]
readme = "README.md"

# name 변경 이유 
# Poetry는 이 이름을 사용하여 패키지를 식별합니다
# 현재 코드에서 lab_competitive_analysis를 임포트하려고 하므로, 패키지 이름이 이와 일치해야 합니다
# "poetry"라는 이름은 너무 일반적이며, 실제 프로젝트와 관련이 없습니다

[tool.poetry.dependencies]
python = "^3.13"
kedro = "^0.19.10"
ipython = "^8.30.0"
kedro-viz = "^10.1.0"
kedro-datasets = "^6.0.0"
matplotlib = "^3.10.0"
seaborn = "^0.13.2"
scikit-learn = "^1.6.0"
openpyxl = "^3.1.5"
pytest = "^8.3.4"
pytest-cov = "^5.0.0"
pytest-mock = "^3.14.0"
pyarrow = "^18.1.0"
fastparquet = "^2024.11.0"
ipykernel = "^6.29.5"
langchain = "^0.3.14"
langchain-community = "^0.3.14"
bs4 = "^0.0.2"
requests = "^2.32.3"
beautifulsoup4 = "^4.12.3"
html2text = "^2024.2.26"
pydantic = "^2.10.5"
langchain-openai = "^0.3.0"
trulens-core = "^1.3.0"
trulens-dashboard = "^1.3.0"
trulens-feedback = "^1.3.0"
trulens-providers-openai = "^1.3.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.kedro]
package_name = "lab_competitive_analysis" #subdirectory_name
project_name = "LAB-COMPETITIVE-ANALYSIS" #rootdirectory_name
kedro_init_version = "0.19.7" #same kedro version in depandency
tools = [ "Data Structure", "Kedro Viz",]
example_pipeline = "True"
source_dir = "lab_competitive_analysis" #subdirectory_name


[tool.kedro_telemetry]
project_id = "935d9e44bd764fc0960d51b67e46ede7"
